grsec: Apply grsecurity-3.0-3.2.53-201312151209.patch
commit cd84f8c580fc31535dcabea0ba8b562697a43d6e
Merge: 33d1bff 9c07653
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 11:58:18 2013 -0500

    Merge branch 'pax-stable2' into grsec-stable2

commit 9c07653f39572a5fc12e025a8ae170a9d0fe3b18
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 11:58:01 2013 -0500

    compile fix

 arch/x86/kernel/entry_64.S |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

commit 33d1bff83bef36441aac0cfd8a6b409d71f0d876
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 10:54:40 2013 -0500

    Backport upstream commit:
    http://lkml.indiana.edu/hypermail/linux/kernel/1312.1/04774.html

    This overflow is harmless except to think about, but it's best
    to fix it. If userspace does a giant read from /dev/urandom,
    bigger than INT_MAX, then that size gets passed straight
    through extract_entropy_user and xfer_secondary_pool to
    _xfer_secondary_pool as nbytes, and we would store it into
    bytes, which is an int. The result could be negative.

    The consequence is pretty small -- we would pull only the minimum
    amount of entropy, rather than as much as we could up to the size
    of the output pool, and this is urandom so that's fine. But the
    code is a little easier to read if we make it clear that overflow
    isn't an issue. Also we might be less likely to make mistakes like
    the one fixed in the previous commit.

    As a bonus, give a name to the minimum number of bytes to pull,
    which we use twice.


 drivers/char/random.c |   11 +++++------
 1 files changed, 5 insertions(+), 6 deletions(-)

commit 092cfa5d608aff41b9db65f39404a21bb864f86e
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 10:49:31 2013 -0500

    Backport of commit:
    http://lkml.indiana.edu/hypermail/linux/kernel/1312.1/04773.html
    Negative numbers and size_t don't mix. When the total entropy
    available was less than 'reserved', we would fail to enforce any limit
    at all. Fix that. We never care how negative have_bytes - reserved
    is, so just flatten it to zero if negative.

    This behavior entered in 987cd8c30 "random: simplify accounting code"
    a few commits ago. Before that, for a long time we would compare
    have_bytes - reserved (or equivalent) to ibytes or store it into ibytes,
    but only inside a condition that guaranteed it wasn't negative.

    ---
    drivers/char/random.c | 3 ++-
    1 file changed, 2 insertions(+), 1 deletion(-)

 drivers/char/random.c |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

commit b75e5e057346017f9a71f63fcc4d8f7bda13567b
Merge: f3e7ef6 33db4e5
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 10:46:17 2013 -0500

    Merge branch 'pax-stable2' into grsec-stable2

    Conflicts:
    	include/linux/random.h

commit 33db4e57263f16ed44ff12c370b1a1a0c97cfd6e
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Dec 15 10:44:24 2013 -0500

    Update to pax-linux-3.2.53-test130.patch:
    - Emese fixed a bug in the size overflow plugin resulting in false positives on downcasts from 64 bit variables on i386, reported by Huub Reuver
    - added a bit more amd64 kernel page table hardening
    - fixed bad interactions between the KERNEXEC plugin and some gcc features, reported by Amadeusz SÅ‚awiÅ„ski (https://bugs.gentoo.org/show_bug.cgi?id=487938)
      - the mask register has been changed from r10 (used by DRAP) to r12
      - all kernel entry points now allocate a full pt_regs area (it required some non-trivial surgery, some fallout is possible)
    - fixed 32 bit apps executing certain 64 bit ones, reported by Ronny Meeus
    - updated size oveflow plugin from Emese, fixes some false positives reported by Tim Harman and Huub Reuver

 arch/x86/crypto/aes-x86_64-asm_64.S          |    2 +-
 arch/x86/crypto/aesni-intel_asm.S            |  124 +++++++-------
 arch/x86/crypto/blowfish-x86_64-asm_64.S     |   12 +-
 arch/x86/crypto/salsa20-x86_64-asm_64.S      |    2 +-
 arch/x86/crypto/sha1_ssse3_asm.S             |   10 +-
 arch/x86/crypto/twofish-x86_64-asm_64-3way.S |    6 +-
 arch/x86/crypto/twofish-x86_64-asm_64.S      |    4 +-
 arch/x86/ia32/ia32entry.S                    |   16 +-
 arch/x86/include/asm/alternative-asm.h       |    6 +-
 arch/x86/include/asm/atomic64_32.h           |    2 +-
 arch/x86/include/asm/atomic64_64.h           |    2 +-
 arch/x86/include/asm/bitops.h                |    4 +-
 arch/x86/include/asm/calling.h               |  118 +++++++------
 arch/x86/include/asm/ptrace-abi.h            |    1 -
 arch/x86/kernel/entry_64.S                   |   71 ++++----
 arch/x86/kernel/head_64.S                    |   18 ++-
 arch/x86/lib/copy_page_64.S                  |   20 +-
 arch/x86/lib/copy_user_64.S                  |    8 +-
 arch/x86/lib/copy_user_nocache_64.S          |    8 +-
 arch/x86/lib/csum-copy_64.S                  |   18 +-
 arch/x86/lib/memcpy_64.S                     |   12 +-
 arch/x86/lib/memmove_64.S                    |   32 ++--
 arch/x86/lib/memset_64.S                     |    4 +-
 arch/x86/lib/msr-reg.S                       |   16 +-
 arch/x86/lib/thunk_64.S                      |   10 +-
 arch/x86/mm/gup.c                            |    4 +
 drivers/usb/core/devio.c                     |    6 +-
 fs/binfmt_elf.c                              |   17 +-
 fs/cifs/file.c                               |   10 +-
 include/asm-generic/bitops/__fls.h           |    2 +-
 include/asm-generic/bitops/fls.h             |    2 +-
 include/asm-generic/bitops/fls64.h           |    4 +-
 include/linux/bitops.h                       |    2 +-
 include/linux/random.h                       |    6 +-
 include/linux/vmstat.h                       |    4 +-
 tools/gcc/kernexec_plugin.c                  |   29 ++--
 tools/gcc/size_overflow_hash.data            |    7 -
 tools/gcc/size_overflow_plugin.c             |  239 ++++++++++++++++++--------
 38 files changed, 484 insertions(+), 374 deletions(-)

Signature-tree: e371ef97780beac6170fcdf0d3812c3e613cbf8a

+++ drivers/usb/core/devio.c
@@ -147,7 +147,7 @@ static ssize_t usbdev_read(struct file *file, char __user *buf, size_t nbytes,
 	struct dev_state *ps = file->private_data;
 	struct usb_device *dev = ps->dev;
 	ssize_t ret = 0;
-	unsigned len;
+	size_t len;
 	loff_t pos;
 	int i;
 
@@ -189,13 +189,13 @@ static ssize_t usbdev_read(struct file *file, char __user *buf, size_t nbytes,
 	for (i = 0; nbytes && i < dev->descriptor.bNumConfigurations; i++) {
 		struct usb_config_descriptor *config =
 			(struct usb_config_descriptor *)dev->rawdescriptors[i];
-		unsigned int length = le16_to_cpu(config->wTotalLength);
+		size_t length = le16_to_cpu(config->wTotalLength);
 
 		if (*ppos < pos + length) {
 
 			/* The descriptor may claim to be longer than it
 			 * really is.  Here is the actual allocated length. */
-			unsigned alloclen =
+			size_t alloclen =
 				le16_to_cpu(dev->config[i].desc.wTotalLength);
 
 			len = length - (*ppos - pos);
